requirements.txt
streamlit
pandas
numpy
torch
transformers
sentence-transformers
faiss-cpu
